requirements.txt
django==2.0.7
pySerial==3.4
